entityshape-66 Rename the src directory to entityshape to facilitate using the code as a python module
