+++ entityshape/compareshape.py
