+++ entityshape/utilities.py
