+++ tests/test_v1.py
@@ -5,7 +5,7 @@ import unittest
 
 import requests
 
-from src.app import app
+from entityshape.app import app
 
 
 class V1Tests(unittest.TestCase):
